pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
aiomysql = "^0.2.0"
alembic = "^1.13.3"
pymysql = "^1.1.1"
openai = "^1.53.0"
pandas = "^2.2.3"
numpy = "^2.1.2"
matplotlib = "^3.9.2"
requests = "^2.32.3"
python-multipart = "^0.0.16"
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pymupdf = "^1.25.1"
aiofiles = "^24.1.0"
pytest = "^8.3.4"
fitz = "^0.0.1.dev2"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

